make 64 bit code work

+++ src/stack.h
@@ -14,7 +14,7 @@ char stack_full(stack *p);
 void stack_push(stack *p, unsigned char x);
 unsigned char stack_peek(stack *p);
 unsigned char stack_pop(stack *p);
-void stack_pop_width(stack *p, int width, unsigned long long int *bytes);
+void stack_pop_width(stack *p, int width, unsigned long *bytes);
 void stack_dump(stack *p);
 
 #endif
